+++ src/semantic/src/identifier-case/index.ts
@@ -1,6 +1,6 @@
 export * from "./conflict-severity.js";
 export * from "./environment.js";
-export * from "./identifier-case-context.js";
+export * from "./identifier-case-helpers.js";
 export * from "./identifier-case-report.js";
 export * from "./identifier-case-utils.js";
 export * from "./option-store.js";
@@ -11,11 +11,9 @@ export * from "./plan-service.js";
 // Local plan exports are intentionally not re-exported to avoid duplicating
 // public APIs that 'plan-service' already exposes.
 export * from "./asset-renames/index.js";
-export * from "./fs-facade.js";
 export * from "./planning-helpers.js";
 export * from "./project-index-bootstrap.js";
 export * from "./project-index-gateway.js";
 // plan-state exports are intentionally kept internal to avoid duplicating
 // names that the public service facade already provides. Import specific
 // helpers from plan-state only when needed by internal modules.
-export * from "./logger.js";
