+++ src/semantic/test/identifier-case-asset-rename-executor.test.ts
@@ -4,7 +4,7 @@ import path from "node:path";
 import { describe, it } from "node:test";
 
 import { __private__, createAssetRenameExecutor } from "../src/identifier-case/asset-renames/executor.js";
-import { DEFAULT_WRITE_ACCESS_MODE } from "../src/identifier-case/fs-facade.js";
+import { DEFAULT_WRITE_ACCESS_MODE } from "../src/identifier-case/identifier-case-helpers.js";
 
 const { ensureWritableDirectory, ensureWritableFile, readJsonFile } = __private__;
