+++ src/semantic/test/identifier-case-report.test.ts
@@ -7,7 +7,7 @@ import { describe, it } from "node:test";
 import {
     clearIdentifierCaseDryRunContexts,
     setIdentifierCaseDryRunContext
-} from "../src/identifier-case/identifier-case-context.js";
+} from "../src/identifier-case/identifier-case-helpers.js";
 import {
     maybeReportIdentifierCaseDryRun,
     summarizeIdentifierCasePlan
